Split parallel_utils.h to header/source, and make bits usable from multiple calling objects
